requirements.txt
# Core Dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning
torch>=1.9.0
tensorflow>=2.6.0
stable-baselines3>=1.0
gym>=0.18.0

# Computer Vision
opencv-python>=4.5.0
scikit-image>=0.18.0

# Traffic Simulation
sumo>=1.9.0  # You might need to install this separately, see SUMO documentation

# Dashboard and Visualization
streamlit>=1.0.0
plotly>=5.0.0
flask>=2.0.0

# Utilities
tqdm>=4.60.0
pyyaml>=5.4.0
